SASS-2199 - Merge employment summary views.
