+++ app/views/employment/EmploymentSummaryView.scala.html
@@ -1,0 +1,135 @@
+@*
+ * Copyright 2022 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *@
+
+@import views.html.templates.Layout
+@import views.html.templates.helpers.Button
+@import views.html.templates.helpers.Heading
+@import utils.ViewUtils.summaryListRow
+@import models.employment.EmploymentSource
+@import controllers.employment.routes.EmployerInformationController
+@import controllers.employment.routes.EmployerNameController
+@import controllers.employment.routes.EmploymentSummaryController
+@import controllers.employment.routes.RemoveEmploymentController
+@import controllers.expenses.routes.CheckEmploymentExpensesController
+@import controllers.expenses.routes.EmploymentExpensesController
+@import common.UUID
+
+@this(
+    layout: Layout,
+    button: Button,
+    heading: Heading,
+    govukSummaryList: GovukSummaryList,
+    govukInsetText: GovukInsetText,
+    govUkButton: GovukButton
+)
+@(taxYear: Int, employments: Seq[EmploymentSource], expensesExist: Boolean, isInYear: Boolean)(implicit user: User[_], messages: Messages, appConfig: AppConfig)
+
+@headingForUse = @{
+    messages("employment.employmentSummary.heading")
+}
+
+@layout(pageTitle = headingForUse, taxYear = Some(taxYear), isAgent = user.isAgent) {
+
+    @heading(headingForUse, Some(messages("payeEmployment.caption", (taxYear - 1).toString, taxYear.toString)), "govuk-!-margin-bottom-2")
+
+    @if(isInYear) {
+        @govukInsetText(InsetText(
+            content = Text(messages(s"employment.employmentInfo.p1.${if(user.isAgent) "agent" else "individual"}", taxYear.toString))
+        ))
+    }
+
+    @{
+        val headingMessage = if (employments.size > 1) messages("employment.employmentSummary.employers") else messages("common.employer")
+        <h2 class="govuk-heading-m govuk-!-margin-top-6">{headingMessage}</h2>
+    }
+
+    @if(appConfig.studentLoansEnabled) {
+        <p class="govuk-body">@messages(s"employment.employmentSummary.content.${if(user.isAgent) "agent" else "individual"}")</p>
+    } else {
+        <p class="govuk-body">@messages(s"employment.employmentSummary.content.studentLoansUnreleased.${if(user.isAgent) "agent" else "individual"}")</p>
+    }
+
+    @{
+        def mapToSummaryListRow(employmentSource: EmploymentSource, isInYear: Boolean): SummaryListRow = {
+            val message = if (isInYear) messages("common.view") else messages("common.change")
+            val actions = if (isInYear) Seq() else Seq((RemoveEmploymentController.show(taxYear, employmentSource.employmentId), messages("common.remove"), Some(messages("common.remove") + s" ${employmentSource.employerName}")))
+
+            summaryListRow(
+                key = HtmlContent(employmentSource.employerName),
+                value = HtmlContent(s"""<a class="govuk-link" href="${EmployerInformationController.show(taxYear, employmentSource.employmentId)}"><span aria-hidden="true">${message}</span><span class="govuk-visually-hidden">${message} ${employmentSource.employerName}</span></a>"""),
+                keyClasses = "govuk-!-width-two-thirds",
+                valueClasses = "govuk-!-width-one-third right-align",
+                actions = actions
+            )
+        }
+
+        govukSummaryList(SummaryList(employments.map(employmentSource => mapToSummaryListRow(employmentSource, isInYear))))
+    }
+
+    @if(!isInYear) {
+        <p class="govuk-body"><a class="govuk-link" href="@EmployerNameController.show(taxYear, UUID.randomUUID)">@messages("employment.employmentSummary.addAnotherEmployer")</a></p>
+    }
+
+    <h2 class="govuk-label--m govuk-!-margin-top-9">@messages("employment.expenses.label")</h2>
+    @if(expensesExist){
+        <p class="govuk-body">@messages(s"employment.expenses.subheading")</p>
+    }
+
+    @if(expensesExist) {
+        @{
+            val message = if (isInYear) messages("common.view") else messages("common.change")
+            val actions = if (isInYear) Seq() else Seq((CheckEmploymentExpensesController.show(taxYear), messages("common.remove"), Some(messages("common.remove") + " " + messages("employment.expenses.label"))))
+
+            val expensesRow = summaryListRow(
+                key = HtmlContent(messages("employment.expenses.label")),
+                value = HtmlContent(s"""<a class="govuk-link" href="${CheckEmploymentExpensesController.show(taxYear)}"><span aria-hidden="true">${message}</span><span class="govuk-visually-hidden">${message} ${messages("employment.expenses.label")}</span></a>"""),
+                keyClasses = "govuk-!-width-two-thirds",
+                valueClasses = "govuk-!-width-one-third right-align",
+                actions = actions
+            )
+
+            govukSummaryList(SummaryList(Seq(expensesRow)))
+        }
+    } else {
+        @if(isInYear) {
+            <p class="govuk-body">@messages(s"employment.employmentSummary.cannotAdd.${if(user.isAgent) "agent" else "individual"}", taxYear.toString)</p>
+        } else {
+            @{
+                val row = summaryListRow(
+                    key = HtmlContent(messages("employment.employmentSummary.noExpensesAdded")),
+                    value = HtmlContent(s"""<a class="govuk-link" href="${EmploymentExpensesController.show(taxYear)}">${messages("common.add")}</a>"""),
+                    keyClasses = "govuk-!-width-two-thirds",
+                    valueClasses = "govuk-!-width-one-third right-align",
+                    actions = Seq()
+                )
+
+                govukSummaryList(SummaryList(Seq(row)))
+            }
+        }
+    }
+
+    @govUkButton(Button(
+        attributes = Map("id" -> "returnToOverviewPageBtn"),
+        preventDoubleClick = true,
+        content = Text(messages("common.return.overview.buttonText")),
+        classes = "govuk-button govuk-!-margin-top-6",
+        href = Some(appConfig.incomeTaxSubmissionOverviewUrl(taxYear))
+    ))
+}
+
+@{
+// $COVERAGE-OFF$
+}
