--- app/views/employment/SingleEmploymentSummaryViewEOY.scala.html
@@ -1,126 +1,0 @@
-@*
- * Copyright 2022 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import views.html.templates.Layout
-@import views.html.templates.helpers.Button
-@import views.html.templates.helpers.Heading
-@import utils.ViewUtils.summaryListRow
-@import models.employment.EmploymentSource
-@import controllers.employment.routes.{EmploymentSummaryController, EmployerNameController}
-@import controllers.expenses.routes.{CheckEmploymentExpensesController, EmploymentExpensesController}
-@import common.UUID
-
-@this(
-    layout: Layout,
-    button: Button,
-    heading: Heading,
-    govukSummaryList: GovukSummaryList,
-    govukInsetText: GovukInsetText,
-    govUkButton: GovukButton
-)
-@(taxYear: Int, employment: EmploymentSource, expensesExist: Boolean)(implicit user: User[_], messages: Messages, appConfig: AppConfig)
-
-@headingForUse = @{
-    messages(s"employment.employmentSummary.heading")
-}
-
-@secondHeader = @{
-    messages("employment.addAnother")
-}
-
-@hint = {
-    @messages(s"employment.addAnother.hint.${if(user.isAgent) "agent" else "individual"}")
-}
-
-@layout(pageTitle = headingForUse, taxYear = Some(taxYear), isAgent = user.isAgent) {
-
-    @heading(headingForUse, Some(messages("payeEmployment.caption", (taxYear - 1).toString, taxYear.toString)), "govuk-!-margin-bottom-2")
-
-    <h2 class="govuk-heading-m govuk-!-margin-top-6">@messages("common.employer")</h2>
-
-    <p class="govuk-body">@messages(s"employment.employmentSummary.content.${if(user.isAgent) "agent" else "individual"}")</p>
-
-    @{
-        val row = summaryListRow(
-                key = HtmlContent(employment.employerName),
-                value = HtmlContent(s"""<a class="govuk-link" href="${controllers.employment.routes.EmployerInformationController.show(taxYear, employment.employmentId)}"><span aria-hidden="true">${messages("common.change")}</span><span class="govuk-visually-hidden">${messages("common.change")} ${employment.employerName}</span></a>"""),
-                keyClasses = "govuk-!-width-two-thirds",
-                valueClasses = "govuk-!-width-one-third right-align",
-                actions = Seq((controllers.employment.routes.RemoveEmploymentController.show(taxYear, employment.employmentId), messages("common.remove"), Some(messages("common.remove") + s" ${employment.employerName}") ))
-            )
-
-        govukSummaryList(
-            SummaryList(
-                Seq(row)
-            )
-        )
-    }
-
-
-    <p class="govuk-body">
-        <a class="govuk-link" href="@EmployerNameController.show(taxYear, UUID.randomUUID)">
-            @messages("employment.employmentSummary.addAnotherEmployer")
-        </a>
-    </p>
-
-    <h2 class="govuk-label--m govuk-!-margin-top-9">@messages("employment.expenses.label")</h2>
-    @if(expensesExist){<p class="govuk-body">@messages(s"employment.expenses.subheading")</p>}
-
-    @if(expensesExist){
-        @{
-            val row = summaryListRow(
-                key = HtmlContent(messages("employment.expenses.label")),
-                value = HtmlContent(s"""<a class="govuk-link" href="${CheckEmploymentExpensesController.show(taxYear)}"><span aria-hidden="true">${messages("common.change")}</span><span class="govuk-visually-hidden">${messages("common.change")} ${messages("employment.expenses.label")}</span></a>"""),
-                keyClasses = "govuk-!-width-two-thirds",
-                valueClasses = "govuk-!-width-one-third right-align",
-                actions = Seq((CheckEmploymentExpensesController.show(taxYear), messages("common.remove"), Some(messages("common.remove") + " " + messages("employment.expenses.label"))))
-            )
-
-            govukSummaryList(
-            SummaryList(
-            Seq(row)
-            ))
-        }
-    } else {
-        @{
-            val row = summaryListRow(
-                key = HtmlContent(messages("employment.employmentSummary.noExpensesAdded")),
-                value = HtmlContent(s"""<a class="govuk-link" href="${EmploymentExpensesController.show(taxYear)}">${messages("common.add")}</a>"""),
-                keyClasses = "govuk-!-width-two-thirds",
-                valueClasses = "govuk-!-width-one-third right-align",
-                actions = Seq()
-            )
-
-            govukSummaryList(
-            SummaryList(
-            Seq(row)
-            )
-            )
-        }
-    }
-
-    @govUkButton(Button(
-        attributes = Map("id" -> "returnToOverviewPageBtn"),
-        preventDoubleClick = true,
-        content = Text(messages("common.return.overview.buttonText")),
-        classes = "govuk-button govuk-!-margin-top-6",
-        href = Some(appConfig.incomeTaxSubmissionOverviewUrl(taxYear))
-    ))
-}
-
-@{
-// $COVERAGE-OFF$
-}
