+++ it/controllers/employment/EmploymentSummaryControllerISpec.scala
@@ -20,25 +20,28 @@ import builders.models.IncomeTaxUserDataBuilder.anIncomeTaxUserData
 import builders.models.employment.AllEmploymentDataBuilder.anAllEmploymentData
 import builders.models.employment.EmploymentSourceBuilder.anEmploymentSource
 import models.IncomeTaxUserData
+import models.employment.{AllEmploymentData, EmploymentSource}
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Document
 import play.api.http.HeaderNames
-import play.api.http.Status.OK
+import play.api.http.Status.{OK, SEE_OTHER, UNAUTHORIZED}
 import play.api.libs.ws.WSResponse
 import play.api.mvc.Result
 import play.api.test.FakeRequest
 import play.api.test.Helpers.route
-import utils.PageUrls.{checkYourExpensesUrl, claimEmploymentExpensesUrl, employerInformationUrl, employerNameUrlWithoutEmploymentId, employmentSummaryUrl, fullUrl, removeEmploymentUrl}
+import utils.PageUrls.{addEmploymentUrl, checkYourExpensesUrl, claimEmploymentExpensesUrl, employerInformationUrl, employerNameUrlWithoutEmploymentId, employmentSummaryUrl, fullUrl, overviewUrl, removeEmploymentUrl}
 import utils.{EmploymentDatabaseHelper, IntegrationTest, ViewHelpers}
 
 import scala.concurrent.Future
 
-class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with ViewHelpers with EmploymentDatabaseHelper {
+class EmploymentSummaryControllerISpec extends IntegrationTest with ViewHelpers with EmploymentDatabaseHelper {
 
-  private val taxYearEOY: Int = taxYear - 1
+  private val taxYearEOY = taxYear - 1
   private val employmentId = "employmentId"
   private val employmentId2 = "002"
   private val employerName2 = "Ken Bosford"
+  private val empWithNoBenefitsOrExpenses: IncomeTaxUserData = anIncomeTaxUserData.copy(employment = Some(anAllEmploymentData.copy(hmrcExpenses = None,
+    hmrcEmploymentData = Seq(anEmploymentSource.copy(employmentBenefits = None)))))
   private val multipleEmpsWithExpenses: IncomeTaxUserData = anIncomeTaxUserData.copy(employment = Some(anAllEmploymentData.copy(
     hmrcEmploymentData = Seq(anEmploymentSource, anEmploymentSource.copy(employmentId = employmentId2, employerName = employerName2)))))
   private val multipleEmpsWithoutExpenses: IncomeTaxUserData = anIncomeTaxUserData.copy(employment = Some(anAllEmploymentData.copy(
@@ -52,6 +55,9 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
     def yourEmpInfoSelector(id: Int): String = s"#main-content > div > div > p:nth-child($id)"
 
     def employerNameSelector(id: Int): String = s"#main-content > div > div > dl:nth-child(5) > div:nth-child($id) > dt"
+
+    val employerNameSelector = "#main-content > div > div > dl:nth-child(4) > div > dt"
+    val employerNameInYearSelector = "#main-content > div > div > dl:nth-child(5) > div > dt"
 
     def employerNameEOYSelector(id: Int): String = s"#main-content > div > div > dl:nth-child(4) > div:nth-child($id) > dt"
 
@@ -95,6 +101,7 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
     val thisIsATotal: String
     val expenses: String
     val noExpensesAdded: String
+    val employer: String
     val employers: String
     val returnToOverview: String
     val employmentDetails: String
@@ -117,6 +124,7 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
     val thisIsATotal: String = "This is a total of expenses from all employment in the tax year."
     val expenses: String = "Expenses"
     val noExpensesAdded: String = "No expenses added"
+    val employer: String = "Employer"
     val employers: String = "Employers"
     val returnToOverview: String = "Return to overview"
     val employmentDetails: String = "Employment details"
@@ -139,6 +147,7 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
     val thisIsATotal: String = "This is a total of expenses from all employment in the tax year."
     val expenses: String = "Expenses"
     val noExpensesAdded: String = "No expenses added"
+    val employer: String = "Employer"
     val employers: String = "Employers"
     val returnToOverview: String = "Return to overview"
     val employmentDetails: String = "Employment details"
@@ -175,14 +184,12 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
     val cannotAdd: String = s"You cannot add your client’s expenses until 6 April $taxYear."
   }
 
-  val userScenarios: Seq[UserScenario[CommonExpectedResults, SpecificExpectedResults]] = {
-    Seq(
-      UserScenario(isWelsh = false, isAgent = false, CommonExpectedEN, Some(ExpectedIndividualEN)),
-      UserScenario(isWelsh = false, isAgent = true, CommonExpectedEN, Some(ExpectedAgentEN)),
-      UserScenario(isWelsh = true, isAgent = false, CommonExpectedCY, Some(ExpectedIndividualCY)),
-      UserScenario(isWelsh = true, isAgent = true, CommonExpectedCY, Some(ExpectedAgentCY))
-    )
-  }
+  val userScenarios: Seq[UserScenario[CommonExpectedResults, SpecificExpectedResults]] = Seq(
+    UserScenario(isWelsh = false, isAgent = false, CommonExpectedEN, Some(ExpectedIndividualEN)),
+    UserScenario(isWelsh = false, isAgent = true, CommonExpectedEN, Some(ExpectedAgentEN)),
+    UserScenario(isWelsh = true, isAgent = false, CommonExpectedCY, Some(ExpectedIndividualCY)),
+    UserScenario(isWelsh = true, isAgent = true, CommonExpectedCY, Some(ExpectedAgentCY))
+  )
 
   ".show" when {
     import Selectors._
@@ -193,9 +200,164 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
       val specific = user.specificExpectedResults.get
 
       s"language is ${welshTest(user.isWelsh)} and request is from an ${agentTest(user.isAgent)}" should {
+        "return the single employment summary EOY page" when {
+          "there is only one employment and its the EOY with expenses" which {
+            implicit lazy val result: WSResponse = {
+              authoriseAgentOrIndividual(user.isAgent)
+              userDataStub(anIncomeTaxUserData, nino, taxYearEOY)
+              urlGet(fullUrl(employmentSummaryUrl(taxYearEOY)), welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
+            }
+
+            lazy val document = Jsoup.parse(result.body)
+
+            implicit def documentSupplier: () => Document = () => document
+
+            "status OK" in {
+              result.status shouldBe OK
+            }
+
+            welshToggleCheck(user.isWelsh)
+            titleCheck(expectedTitle)
+            h1Check(expectedH1)
+            captionCheck(expectedCaption(taxYearEOY))
+            textOnPageCheck(employer, employersSelector)
+            textOnPageCheck(specific.yourEmpInfo, yourEmpInfoSelector(3))
+            textOnPageCheck(name, employerNameSelector)
+            linkCheck(s"$change$change $name", changeEmployerSelector(1), employerInformationUrl(taxYearEOY, employmentId))
+            linkCheck(s"$remove $remove $name", removeEmployerSelector(1), removeEmploymentUrl(taxYearEOY, employmentId))
+            linkCheck(addAnother, addAnotherSelector, employerNameUrlWithoutEmploymentId(taxYearEOY), isExactUrlMatch = false)
+            textOnPageCheck(expenses, expensesHeadingSelector, "as a heading")
+            textOnPageCheck(expenses, expensesLineSelector, "as a line item")
+            linkCheck(s"$change$change $expenses", changeExpensesSelector, checkYourExpensesUrl(taxYearEOY))
+            linkCheck(s"$remove $remove $expenses", removeExpensesSelector, checkYourExpensesUrl(taxYearEOY))
+            buttonCheck(returnToOverview)
+          }
+
+          "there is only one employment and its the EOY without expenses" which {
+            implicit lazy val result: WSResponse = {
+              authoriseAgentOrIndividual(user.isAgent)
+              userDataStub(anIncomeTaxUserData.copy(employment = Some(anAllEmploymentData.copy(hmrcExpenses = None))), nino, taxYearEOY)
+              urlGet(fullUrl(employmentSummaryUrl(taxYearEOY)), welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
+            }
+
+            lazy val document = Jsoup.parse(result.body)
+
+            implicit def documentSupplier: () => Document = () => document
+
+            "status OK" in {
+              result.status shouldBe OK
+            }
+
+            welshToggleCheck(user.isWelsh)
+            titleCheck(expectedTitle)
+            h1Check(expectedH1)
+            captionCheck(expectedCaption(taxYearEOY))
+            textOnPageCheck(employer, employersSelector)
+            textOnPageCheck(specific.yourEmpInfo, yourEmpInfoSelector(3))
+            textOnPageCheck(name, employerNameSelector)
+            linkCheck(s"$change$change $name", changeEmployerSelector(1), employerInformationUrl(taxYearEOY, employmentId))
+            linkCheck(s"$remove $remove $name", removeEmployerSelector(1), removeEmploymentUrl(taxYearEOY, employmentId))
+            linkCheck(addAnother, addAnotherSelector, employerNameUrlWithoutEmploymentId(taxYearEOY), isExactUrlMatch = false)
+            textOnPageCheck(expenses, expensesHeadingSelector, "as a heading")
+            textOnPageCheck(noExpensesAdded, noExpensesAddedSelector)
+            linkCheck(add, addSelector, claimEmploymentExpensesUrl(taxYearEOY))
+            buttonCheck(returnToOverview)
+          }
+        }
+
+        "return the single employment summary in year page" when {
+          "there is only one employment and its in year with expenses" which {
+            implicit lazy val result: WSResponse = {
+              authoriseAgentOrIndividual(user.isAgent)
+              userDataStub(anIncomeTaxUserData, nino, taxYear)
+              urlGet(fullUrl(employmentSummaryUrl(taxYear)), welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYear)))
+            }
+
+            lazy val document = Jsoup.parse(result.body)
+
+            implicit def documentSupplier: () => Document = () => document
+
+            "status OK" in {
+              result.status shouldBe OK
+            }
+
+            welshToggleCheck(user.isWelsh)
+            titleCheck(expectedTitle)
+            h1Check(expectedH1)
+            captionCheck(expectedCaption(taxYear))
+            textOnPageCheck(specific.cannotUpdateInfo, cannotUpdateInfoSelector)
+            textOnPageCheck(employer, employersSelector)
+            textOnPageCheck(specific.yourEmpInfo, yourEmpInfoSelector(4))
+            textOnPageCheck(name, employerNameInYearSelector)
+            linkCheck(s"$view$view $name", viewEmployerSelector(1), employerInformationUrl(taxYear, employmentId))
+            textOnPageCheck(expenses, expensesHeadingSelector, "as a heading")
+            textOnPageCheck(thisIsATotal, thisIsATotalSelector)
+            textOnPageCheck(expenses, expensesLineSelector, "as a line item")
+            linkCheck(s"$view$view $expenses", viewExpensesSelector, checkYourExpensesUrl(taxYear))
+            buttonCheck(returnToOverview)
+          }
+
+          "there is only one employment and its in year without expenses or benefits" which {
+            implicit lazy val result: WSResponse = {
+              authoriseAgentOrIndividual(user.isAgent)
+              userDataStub(empWithNoBenefitsOrExpenses, nino, taxYear)
+              urlGet(fullUrl(employmentSummaryUrl(taxYear)), welsh = user.isWelsh, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYear)))
+            }
+
+            lazy val document = Jsoup.parse(result.body)
+
+            implicit def documentSupplier: () => Document = () => document
+
+            "status OK" in {
+              result.status shouldBe OK
+            }
+
+            welshToggleCheck(user.isWelsh)
+            titleCheck(expectedTitle)
+            h1Check(expectedH1)
+            captionCheck(expectedCaption(taxYear))
+            textOnPageCheck(specific.cannotUpdateInfo, cannotUpdateInfoSelector)
+            textOnPageCheck(employer, employersSelector)
+            textOnPageCheck(specific.yourEmpInfo, yourEmpInfoSelector(4))
+            textOnPageCheck(name, employerNameInYearSelector)
+            linkCheck(s"$view$view $name", viewEmployerSelector(1), employerInformationUrl(taxYear, employmentId))
+            textOnPageCheck(expenses, expensesHeadingSelector, "as a heading")
+            textOnPageCheck(specific.cannotAdd, cannotAddSelector)
+            buttonCheck(returnToOverview)
+          }
+        }
+
+        "return the single employment summary in year page without references to student loans" when {
+          "the student loans feature switch is false" which {
+            val headers = if (user.isWelsh) {
+              Seq(HeaderNames.COOKIE -> playSessionCookies(taxYear), HeaderNames.ACCEPT_LANGUAGE -> "cy")
+            } else {
+              Seq(HeaderNames.COOKIE -> playSessionCookies(taxYear))
+            }
+
+            val request = FakeRequest("GET", employmentSummaryUrl(taxYear)).withHeaders(headers: _*)
+
+            lazy val result: Future[Result] = {
+              authoriseAgentOrIndividual(user.isAgent)
+              userDataStub(anIncomeTaxUserData, nino, taxYear)
+              route(appWithFeatureSwitchesOff, request, "{}").get
+            }
+
+            implicit def document: () => Document = () => Jsoup.parse(bodyOf(result))
+
+            "status OK" in {
+              status(result) shouldBe OK
+            }
+
+            welshToggleCheck(user.isWelsh)
+            titleCheck(expectedTitle)
+            h1Check(expectedH1)
+            textOnPageCheck(specific.yourEmpInfoStudentLoansUnreleased, yourEmpInfoSelector(4))
+            buttonCheck(returnToOverview)
+          }
+        }
 
         "return the multiple employment summary in year page" when {
-
           "there are 2 employments and its in year showing 2 employments with expenses" which {
             implicit lazy val result: WSResponse = {
               authoriseAgentOrIndividual(user.isAgent)
@@ -204,7 +366,8 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
             }
 
             lazy val document = Jsoup.parse(result.body)
-          implicit def documentSupplier: () => Document = () => document
+
+            implicit def documentSupplier: () => Document = () => document
 
             "status OK" in {
               result.status shouldBe OK
@@ -236,7 +399,8 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
             }
 
             lazy val document = Jsoup.parse(result.body)
-          implicit def documentSupplier: () => Document = () => document
+
+            implicit def documentSupplier: () => Document = () => document
 
             "status OK" in {
               result.status shouldBe OK
@@ -257,11 +421,9 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
             textOnPageCheck(specific.cannotAdd, cannotAddSelector)
             buttonCheck(returnToOverview)
           }
-
         }
 
         "return the multiple employment summary EOY page" when {
-
           "there are 2 employments and its EOY showing 2 employments with expenses" which {
             implicit lazy val result: WSResponse = {
               authoriseAgentOrIndividual(user.isAgent)
@@ -270,7 +432,8 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
             }
 
             lazy val document = Jsoup.parse(result.body)
-          implicit def documentSupplier: () => Document = () => document
+
+            implicit def documentSupplier: () => Document = () => document
 
             "status OK" in {
               result.status shouldBe OK
@@ -305,7 +468,8 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
             }
 
             lazy val document = Jsoup.parse(result.body)
-          implicit def documentSupplier: () => Document = () => document
+
+            implicit def documentSupplier: () => Document = () => document
 
             "status OK" in {
               result.status shouldBe OK
@@ -332,9 +496,8 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
         }
 
         "return the multiple employment summary EOY page without references to student loans" when {
-
           "the student loans feature switch is false" which {
-            val headers = if (user.isWelsh){
+            val headers = if (user.isWelsh) {
               Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY), HeaderNames.ACCEPT_LANGUAGE -> "cy")
             } else {
               Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY))
@@ -361,6 +524,57 @@ class MultipleEmploymentSummaryControllerISpec extends IntegrationTest with View
             buttonCheck(returnToOverview)
           }
         }
+      }
+    }
+
+    "redirect when there is employment data returned but no hmrc employment data" which {
+      lazy val result: WSResponse = {
+        authoriseAgentOrIndividual(isAgent = true)
+        userDataStub(IncomeTaxUserData(Some(AllEmploymentData(Seq(), None, Seq(anEmploymentSource), None))), nino, taxYear)
+        urlGet(fullUrl(employmentSummaryUrl(taxYear)), follow = false, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYear)))
+      }
+
+      "status OK" in {
+        result.status shouldBe SEE_OTHER
+      }
+    }
+
+    "redirect the User to the Overview page no data in session" which {
+      lazy val result: WSResponse = {
+        authoriseAgentOrIndividual(isAgent = true)
+        userDataStub(IncomeTaxUserData(), nino, taxYear)
+        urlGet(fullUrl(employmentSummaryUrl(taxYear)), follow = false, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYear)))
+      }
+
+      "has an SEE_OTHER(303) status" in {
+        result.status shouldBe SEE_OTHER
+        result.header(HeaderNames.LOCATION).contains(overviewUrl(taxYear)) shouldBe true
+      }
+
+    }
+
+    "redirect the user to the Add Employment page when no data is in session EOY" which {
+      lazy val result: WSResponse = {
+        authoriseAgentOrIndividual(isAgent = true)
+        val employmentSources = Seq(EmploymentSource(employmentId = "001", employerName = "maggie", None, None, None, None, dateIgnored = Some("2020-03-11"), None, None, None))
+        userDataStub(IncomeTaxUserData(Some(anAllEmploymentData.copy(hmrcEmploymentData = employmentSources))), nino, taxYearEOY)
+        urlGet(s"$appUrl/$taxYearEOY/employment-summary", follow = false, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
+      }
+
+      "has an SEE_OTHER(303) status" in {
+        result.status shouldBe SEE_OTHER
+        result.header(HeaderNames.LOCATION).contains(addEmploymentUrl(taxYearEOY)) shouldBe true
+      }
+
+    }
+
+    "returns an action when auth call fails" which {
+      lazy val result: WSResponse = {
+        unauthorisedAgentOrIndividual(isAgent = true)
+        urlGet(fullUrl(employmentSummaryUrl(taxYear)))
+      }
+      "has an UNAUTHORIZED(401) status" in {
+        result.status shouldBe UNAUTHORIZED
       }
     }
   }
